Fixed polygon update

+++ PPNetLib/Contracts/UpdatePointReq.cs
@@ -9,8 +9,8 @@ namespace PPNetLib.Contracts
     public class UpdatePointReq: Packet
     {
         [ProtoMember(1)] public int PointId { get; set; }
-        [ProtoMember(2)] public int Lat { get; set; }
-        [ProtoMember(3)] public int Lng { get; set; }
+        [ProtoMember(2)] public double Lat { get; set; }
+        [ProtoMember(3)] public double Lng { get; set; }
 
         public UpdatePointReq()
         {
